refactor - theme data service to own dir
