+++ src/app/app.component.html
@@ -1,4 +1,4 @@
-<app-navbar class="test"></app-navbar>
+<app-navbar></app-navbar>
 <app-query-container></app-query-container>
 
 <router-outlet />
