+++ src/app/components/search-bar/search-bar.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from "@angular/core";
-import { ThemeDataService } from "../../theme-data.service";
+import { ThemeDataService } from "../../services/theme-data/theme-data.service";
 
 @Component({
   selector: "app-search-bar",
@@ -24,3 +24,5 @@ export class SearchBarComponent implements OnInit {
     );
   }
 }
+
+
